+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.0.5] - WIP
+
+### Added
+
+- The support for stubbing constants (nepalez)
+
+  ```yaml
+  # Stub constant TIMEOUT_SEC to 10
+  ---
+  - const: TIMEOUT_SEC
+    value: 10
+  ```
+
 ## [0.0.4] - [2018-05-22]
 
 ### Added